pyproject.toml
[project]
name = "to-do-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apscheduler>=3.11.0",
    "discord.py>=2.3.2",
    "pymongo>=4.13.2",
    "python-dotenv>=1.1.1",
]
